requirements.txt
flask
flask-sqlalchemy
flask-bcrypt
python-dotenv
pg8000
flask-session
redis
flask-cors